# backend/internal/httpd/apispec/openapi.yaml
openapi: 3.1.0
info:
  description: Loopback-only HTTP surface served by the Go daemon. Generated from
    Go (code-first) — do not edit by hand; run `go generate ./...`.
  title: Agent Orchestrator HTTP daemon
  version: 0.1.0-route-shell
servers:
- description: Local daemon (loopback only)
  url: http://127.0.0.1:3001
paths:
...
  /api/v1/notifications:
    get:
      operationId: listNotifications
      parameters:
      - description: Notification status filter. Defaults to unread; all includes
          read history.
        in: query
        name: status
        schema:
          description: Notification status filter. Defaults to unread; all includes
            read history.
          enum:
          - unread
          - all
          type: string
      - description: Maximum notifications to return. Defaults to 100.
        in: query
        name: limit
        schema:
          description: Maximum notifications to return. Defaults to 100.
          maximum: 100
          minimum: 1
          type: integer
      - description: Opaque cursor returned by the previous page.
        in: query
        name: cursor
        schema:
          description: Opaque cursor returned by the previous page.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListNotificationsResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: List notification history
      tags:
      - notifications
  /api/v1/notifications/{id}:
    patch:
      operationId: markNotificationRead
      parameters:
      - description: Notification identifier.
        in: path
        name: id
        required: true
        schema:
          description: Notification identifier.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkNotificationReadRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationEnvelope'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Mark a notification read
      tags:
      - notifications
  /api/v1/notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarkAllNotificationsReadResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Mark all unread notifications read
      tags:
      - notifications
  /api/v1/notifications/stream:
    get:
      operationId: streamNotifications
      parameters:
      - description: Optional project id filter for live notifications.
        in: query
        name: projectId
        schema:
          description: Optional project id filter for live notifications.
          type: string
      responses:
        "200":
          content:
            text/event-stream:
              schema:
                type: string
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Stream created notifications
      tags:
      - notifications
...
components:
  schemas:
    APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
      required:
      - error
      - code
      - message
      type: object
...
    ListNotificationsResponse:
      properties:
        nextCursor:
          type: string
        notifications:
          items:
            $ref: '#/components/schemas/NotificationResponse'
          type: array
        unreadCount:
          type: integer
      required:
      - notifications
      - unreadCount
      type: object
...
    MarkAllNotificationsReadResponse:
      properties:
        notifications:
          description: Deprecated compatibility field. Always empty so mark-all responses
            stay bounded.
          items:
            $ref: '#/components/schemas/NotificationResponse'
          type: array
        updatedCount:
          description: Number of notifications changed from unread to read.
          format: int64
          type: integer
      required:
      - notifications
      - updatedCount
      type: object
    MarkNotificationReadRequest:
      properties:
        status:
          description: V1 supports only marking an unread notification read.
          enum:
          - read
          type: string
      required:
      - status
      type: object
...
    NotificationEnvelope:
      properties:
        notification:
          $ref: '#/components/schemas/NotificationResponse'
      required:
      - notification
      type: object
    NotificationResponse:
      properties:
        body:
          type: string
        createdAt:
          format: date-time
          type: string
        id:
          type: string
        prUrl:
          type: string
        projectId:
          type: string
        sessionId:
          type: string
        status:
          enum:
          - unread
          - read
          type: string
        target:
          $ref: '#/components/schemas/NotificationTarget'
        title:
          type: string
        type:
          enum:
          - needs_input
          - ready_to_merge
          - pr_merged
          - pr_closed_unmerged
          type: string
      required:
      - id
      - sessionId
      - projectId
      - prUrl
      - type
      - title
      - body
      - status
      - createdAt
      - target
      type: object
    NotificationTarget:
      properties:
        kind:
          enum:
          - session
          - pr
          type: string
        prUrl:
          type: string
        sessionId:
          type: string
      required:
      - kind
      - sessionId
      type: object
...